+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 
   ### :bug: Bug Fix
 
+- `ui-components/link-preview/vertical-embed`
+  - [#3472](https://github.com/wix/ricos/pull/3472) fix social-embed ui & vertical-embed save logic on mobile
 - `ui-components`
   - [#3471](https://github.com/wix/ricos/pull/3471) fix labeled-toggle text dir